requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy>=1.4.23
asyncpg
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.5
bcrypt>=4.0.1
python-dotenv>=0.19.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
